+++ geoportal/c2cgeoportal_geoportal/views/mapserverproxy.py
@@ -34,12 +34,13 @@ from pyramid.request import Request
 from pyramid.response import Response
 from pyramid.view import view_config
 
+from c2cgeoportal_commons.lib.caching import get_region
+from c2cgeoportal_commons.lib.functionality import get_mapserver_substitution_params
 from c2cgeoportal_commons.lib.url import Url
 from c2cgeoportal_commons.models import main
 from c2cgeoportal_geoportal.lib import get_roles_id, get_roles_name
-from c2cgeoportal_geoportal.lib.caching import Cache, get_region
+from c2cgeoportal_geoportal.lib.caching import Cache
 from c2cgeoportal_geoportal.lib.filter_capabilities import filter_capabilities
-from c2cgeoportal_geoportal.lib.functionality import get_mapserver_substitution_params
 from c2cgeoportal_geoportal.views.ogcproxy import OGCProxy
 
 CACHE_REGION = get_region("std")
